+++ web/src/features/hfiCalculator/pages/HfiCalculatorPage.tsx
@@ -33,7 +33,7 @@ import { DateRange } from 'components/dateRangePicker/types'
 import LiveChangesAlert from 'features/hfiCalculator/components/LiveChangesAlert'
 import { AppDispatch } from 'app/store'
 import HFILoadingDataView from 'features/hfiCalculator/components/HFILoadingDataView'
-import ManageStationsButton from 'features/hfiCalculator/components/ManageStationsButton'
+import ManageStationsButton from 'features/hfiCalculator/components/stationAdmin/ManageStationsButton'
 
 const useStyles = makeStyles(theme => ({
   ...formControlStyles,
